requirements.txt
transformers
datasets
datasketch
redis
kiwipiepy
omegaconf
hydra-core